Add files via upload
silver-asc.lut
This lut is design to be close to a gray ramp but subtly tinted to look good in color. The intensity is linear and has 256 tint variations.

viridis-linearlumin.lut
This is a variant on the mpl-viridis lut that has a linear intensity variation across the ramp so it works better when the image is printed in gray-scale.